channel, opening_control: Make first_blocknum u32
The `new_channel` constructor accepts u32, and the
`get_block_height` function returns u32, so the
extra 32 bits is unuseable anyway.

+++ lightningd/opening_control.c
@@ -49,7 +49,7 @@ struct uncommitted_channel {
 
 	/* Blockheight at creation, scans for funding confirmations
 	 * will start here */
-	u64 first_blocknum;
+	u32 first_blocknum;
 
 	/* These are *not* filled in by new_uncommitted_channel: */
